+++ Windows Universal Samples/CS_sample/MainPage.xaml.cs
@@ -134,10 +134,12 @@ namespace CS_sample
             embeddedControl.AppID = appID;
             embeddedControl.Placements = placement1 + "," + placement2 + "," + placement3;
             embeddedControl.Placement = placement2;
-            embeddedControl.SoundEnabled = false;
+            //embeddedControl.SoundEnabled = false;
 
             embeddedControl.OnAdStart += Embedded_OnAdStart;
             embeddedControl.OnAdEnd += Embedded_OnAdEnd;
+
+            embeddedControl.AdConfig.Volume = 1.0;
 
             var nEmb = await embeddedControl.PlayAdAsync();
         }
@@ -152,6 +154,7 @@ namespace CS_sample
             adConfig.IncentivizedDialogContinueButton = "";
             adConfig.IncentivizedDialogTitle = "";
             adConfig.UserId = "VTest";
+            adConfig.Volume = 1.0;
 
             await sdkInstance.PlayAdAsync(adConfig, placement3);
         }